% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{cis_info}
\alias{cis_info}
\title{Show information about the local CIS dataset}
\usage{
cis_info()
}
\value{
Invisibly, a one-row tibble with local dataset metadata.
}
\description{
Show information about the local CIS dataset
}
