% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cis_open}
\alias{cis_open}
\title{Open the local CIS dataset lazily}
\usage{
cis_open()
}
\value{
An Arrow dataset that can be queried with dplyr verbs.
}
\description{
Open the local CIS dataset lazily
}
\examples{
\dontrun{
cis_open() |>
  dplyr::filter(fecha >= as.Date("2023-01-01")) |>
  dplyr::select(estudio, fecha, genero, edad) |>
  dplyr::collect()
}
}
