% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cis_schema}
\alias{cis_schema}
\title{Return the CIS schema}
\usage{
cis_schema()
}
\value{
A tibble with at least \code{name} and \code{type}.
}
\description{
Uses the local manifest when it includes a schema. Otherwise it falls back to
the Arrow schema in the local Parquet file.
}
