% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cis_studies}
\alias{cis_studies}
\title{List studies available in the local CIS dataset}
\usage{
cis_studies()
}
\value{
A tibble with \code{estudio}, \code{fecha}, \code{anio}, and \code{mes}.
}
\description{
List studies available in the local CIS dataset
}
