% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cislongitudinal-package.R
\docType{package}
\name{cislongitudinal-package}
\alias{cislongitudinal}
\alias{cislongitudinal-package}
\title{cislongitudinal: Access the CIS longitudinal dataset}
\description{
Download, cache, update, inspect, and query the longitudinal CIS dataset
published by Spain Electoral Project as a Parquet file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hmeleiro/cislongitudinal}
  \item \url{https://cislongitudinal.spainelectoralproject.com/}
  \item Report bugs at \url{https://github.com/hmeleiro/cislongitudinal/issues}
}

}
\author{
\strong{Maintainer}: Héctor Meleiro \email{hmeleiros@gmail.com} (\href{https://orcid.org/0009-0003-4468-2197}{ORCID})

}
\keyword{internal}
